default export of currencies, move it to i18n and remove result's slot
