+++ src/i18n/currencies.ts
@@ -2,9 +2,7 @@
  * TODO: This configuration should be part of an intermediate project that would be imported by
  * the X-Archetype.
  */
-const currencies: Record<string, string> = {
+export default {
   EUR: 'i.iii,dd €',
   USD: '$i,iii.dd'
-};
-
-export default currencies;
+} as Record<string, string>;
